Renamed Timer classes. Moved the static methods to the MovedStatics.java class. Additionally removed the dummy parameters from the time related classes.
