+++ src/main/java/com/jagex/runescape/Class51.java
@@ -42,7 +42,7 @@ public class Class51 {
             ActorDefinition.drawMapBack();
             GenericTile.method943(ChatBox.tradeMode, WallDecoration.fontNormal, ChatBox.privateChatMode, ChatBox.publicChatMode);
             Class40_Sub2.method527(Class5.currentTabId, arg0 + 4, Class40_Sub5_Sub11.tabWidgetIds, Class29.tabAreaOverlayWidgetId == -1, -1);
-            Class38.aBoolean893 = true;
+            MovedStatics.aBoolean893 = true;
             Class40_Sub3.aBoolean2026 = true;
             CacheIndex.aBoolean260 = true;
         }
