+++ src/main/java/com/jagex/runescape/Class8_Sub1.java
@@ -128,7 +128,7 @@ public abstract class Class8_Sub1 extends Class8 implements Runnable {
                 if(!aBoolean1820)
                     break;
             }
-            Class43.method890(50L, 48);
+            Class43.sleep(50L);
         }
     }
 
@@ -146,7 +146,7 @@ public abstract class Class8_Sub1 extends Class8 implements Runnable {
                 }
                 method212(System.currentTimeMillis());
             }
-            Class43.method890(5L, -86);
+            Class43.sleep(5L);
         }
     }
