+++ src/main/java/com/jagex/runescape/RSRuntimeException.java
@@ -64,7 +64,7 @@ public class RSRuntimeException extends RuntimeException {
     }
 
     public static void method1057(int arg0) {
-        SceneCluster.aClass38_768.method444(10);
+        SceneCluster.aTimer_768.reset();
         for(int i = 0; i < 32; i++)
             GameShell.aLongArray4[i] = 0L;
         for(int i = 0; i < 32; i++)
