+++ src/main/java/com/jagex/runescape/cache/media/AnimationSequence.java
@@ -70,7 +70,7 @@ public class AnimationSequence extends SubNode {
         int i_15_ = 0x3 & i_14_ >> 6;
         int i_16_ = 0x1f & i_14_;
         if(i_16_ != 10 && i_16_ != 11 && i_16_ != 22)
-            Class38_Sub1.doWalkTo(0, 0, Player.localPlayer.pathY[0], arg0, 1 + i_16_, true, i_15_, 0, Player.localPlayer.pathX[0], arg3, 2);
+            MovedStatics.doWalkTo(0, 0, Player.localPlayer.pathY[0], arg0, 1 + i_16_, true, i_15_, 0, Player.localPlayer.pathX[0], arg3, 2);
         else {
             GameObjectDefinition gameObjectDefinition = GameObjectDefinition.getDefinition(i);
             int i_17_ = gameObjectDefinition.blockingMask;
@@ -85,7 +85,7 @@ public class AnimationSequence extends SubNode {
                 i_18_ = gameObjectDefinition.sizeY;
                 i_19_ = gameObjectDefinition.sizeX;
             }
-            Class38_Sub1.doWalkTo(i_18_, i_17_, Player.localPlayer.pathY[0], arg0, 0, true, 0, i_19_, Player.localPlayer.pathX[0], arg3, 2);
+            MovedStatics.doWalkTo(i_18_, i_17_, Player.localPlayer.pathY[0], arg0, 0, true, 0, i_19_, Player.localPlayer.pathX[0], arg3, 2);
         }
         Class40_Sub11.crossX = Class57.clickX;
         LinkedList.crossType = 2;
