+++ src/main/java/com/jagex/runescape/frame/ScreenMode.java
@@ -1,11 +1,11 @@
 package com.jagex.runescape.frame;
 
 import com.jagex.runescape.Class35;
-import com.jagex.runescape.Class38;
 import com.jagex.runescape.Class40_Sub5_Sub13;
 import com.jagex.runescape.Class51;
 import com.jagex.runescape.media.Rasterizer3D;
 import com.jagex.runescape.scene.Scene;
+import com.jagex.runescape.util.Timer;
 
 import java.awt.*;
