+++ src/main/java/com/jagex/runescape/input/MouseHandler.java
@@ -109,7 +109,7 @@ public class MouseHandler implements MouseListener, MouseMotionListener, FocusLi
             }
         } else {
             if(meta == 1 && ActorDefinition.menuActionRow > 0) {
-                int action = Class38.menuActionTypes[ActorDefinition.menuActionRow - 1];
+                int action = MovedStatics.menuActionTypes[ActorDefinition.menuActionRow - 1];
                 if(action == 53 || action == 25 || action == 55 || action == 48 || action == 24 || action == 52 || action == 6 || action == 31 || action == 43 || action == 11 || action == 19 || action == 1006) {
                     int item = InteractiveObject.firstMenuOperand[ActorDefinition.menuActionRow - 1];
                     int id = Class59.secondMenuOperand[-1 + ActorDefinition.menuActionRow];
